gave it a good spit shine. even sprinkled in a few comments to give it that new code smell

+++ src/model/records/Point.java
@@ -1,4 +1,4 @@
 package model.records;
 
-public record Point(byte x, byte y, byte z) {
+public record Point(int x, int y, int z) {
 }
